Rename form control components name to follow convention
